+++ src/main/java/io/patriot_framework/generator/eventSimulator/eventGenerator/graphFire/Room.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2024 Patriot project
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package io.patriot_framework.generator.eventSimulator.eventGenerator.graphFire;
+
+import java.util.Objects;
+
+public class Room {
+
+    public Room(String name, Integer temp) {
+        this.name = name;
+        this.temp = temp;
+        energy = 100; // todo uzivatel si muze zadat sam
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof Room)) return false;
+        Room vertex = (Room) o;
+        return Objects.equals(name, vertex.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name);
+    }
+
+    public String name;
+    public Integer temp;
+    public Integer energy;
+}
